+++ src/app.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get , Param} from '@nestjs/common';
+import { Controller, Get } from '@nestjs/common';
 import { AppService } from './app.service';
 
 @Controller()
@@ -7,7 +7,6 @@ export class AppController {
 
   @Get()
   getHello(): string {
-    return "Hola Mundo";
+    return 'Hola Mundo';
   }
-
 }
