+++ src/app.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ProductsController } from './controllers/products.controller';
-import { CategoriesController } from './controllers/categories.controller';
-import { ProductsService } from './services/products.service';
+import { ProductsController } from './products/controllers/products.controller';
+import { CategoriesController } from './categories/controllers/categories.controller';
+import { ProductsService } from './products/services/products.service';
+import { Service } from './users/services/.service';
 
 @Module({
   imports: [],
   controllers: [AppController, ProductsController, CategoriesController],
-  providers: [AppService, ProductsService],
+  providers: [AppService, ProductsService, Service],
 })
 export class AppModule {}
