--- src/dtos/product.dto.ts
@@ -1,15 +1,0 @@
-export class CreateProductDTO {
-    readonly name: string;
-    readonly description: string;
-    readonly price: number;
-    readonly image: string;
-    readonly stock: number;
-  }
-
-  export class UpdateProductDto {
-    readonly name?: string;
-    readonly description?: string;
-    readonly price?: number;
-    readonly stock?: number;
-    readonly image?: string;
-  }
