--- src/entities/product.entity.ts
@@ -1,8 +1,0 @@
-export class Product {
-    id: number;
-    name: string;
-    description: string;
-    price: number;
-    stock: number;
-    image: string;
-  }
