+++ src/products/controllers/products.controller.ts
@@ -1,0 +1,46 @@
+import {
+  Controller,
+  Get,
+  Param,
+  Post,
+  Body,
+  Put,
+  Delete,
+  HttpStatus,
+  HttpCode,
+} from '@nestjs/common';
+
+import { ParseIntPipe } from '../../common/parse-int/parse-int.pipe';
+import { ProductsService } from '../services/products.service';
+import { CreateProductDTO, UpdateProductDto } from '../dtos/product.dto';
+
+@Controller('products')
+export class ProductsController {
+  constructor(private productsService: ProductsService) {}
+
+  @Get()
+  getProducts() {
+    return this.productsService.findAll();
+  }
+
+  @Get(':productId')
+  @HttpCode(HttpStatus.ACCEPTED)
+  getOne(@Param('productId', ParseIntPipe) productId: number) {
+    return this.productsService.findOne(productId);
+  }
+
+  @Post()
+  create(@Body() payload: CreateProductDTO) {
+    return this.productsService.create(payload);
+  }
+
+  @Put(':id')
+  update(@Param('id') id: string, @Body() payload: UpdateProductDto) {
+    return this.productsService.update(+id, payload);
+  }
+
+  @Delete(':id')
+  delete(@Param('id') id: string) {
+    return this.productsService.remove(+id);
+  }
+}
