+++ src/products/entities/product.entity.ts
@@ -1,0 +1,8 @@
+export class Product {
+  id: number;
+  name: string;
+  description: string;
+  price: number;
+  stock: number;
+  image: string;
+}
