+++ src/products/services/products.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 
-import { Product } from './../entities/product.entity';
-import {CreateProductDTO, UpdateProductDto } from './../dtos/product.dto';
+import { Product } from '../entities/product.entity';
+import { CreateProductDTO, UpdateProductDto } from '../dtos/product.dto';
 
 @Injectable()
 export class ProductsService {
