+++ src/users/entities/users.entities.ts
@@ -1,0 +1,8 @@
+export class Product {
+  id: number;
+  name: string;
+  pass: string;
+  status: number;
+  type: number;
+  image: string;
+}
